% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_load_radr.R
\name{cn_load_radr}
\alias{cn_load_radr}
\title{CN weather - Downloading RADAR data from http://www.weather.com.cn/}
\usage{
cn_load_radr(
  time = weather2::tool_datetime(end = Sys.time(), duration = "7 days", by = "1 mins")
    \%>\% weather2::tool_datetime_select(by = "min", value = 0),
  list_fail = T,
  dir = getwd(),
  attempt = 2L,
  worker = 1L
)
}
\arguments{
\item{time}{Date/time to download}

\item{list_fail}{List failed-to-download items}

\item{dir}{Directory of downloaded data.}

\item{attempt}{Attempts to be made per file to download}

\item{worker}{Numbers of sessions to be open}
}
\value{
Photos from http://www.weather.com.cn/
}
\description{
CN weather - Downloading RADAR data from http://www.weather.com.cn/
}
\examples{
cn_load_radr()
}
