% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk_load_all.R
\name{hk_load_all}
\alias{hk_load_all}
\title{HK weather - Downloading all information from HKO and SMG}
\usage{
hk_load_all(worker = 1)
}
\arguments{
\item{worker}{Numbers of sessions to be open}
}
\value{

}
\description{
HK weather - Downloading all information from HKO and SMG
}
\examples{
hk_load_all(worker = 1) or hk_load_all(worker = 20)
}
