% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk_load_ltng.R
\name{hk_load_ltng}
\alias{hk_load_ltng}
\title{HK weather - Downloading lighting & radar images from https://www.hko.gov.hk/en/index.html}
\usage{
hk_load_ltng(
  time = weather2::tool_datetime(end = Sys.time(), by = "6 min", duration = "96 hour"),
  type = c("cc", "cg"),
  range = c(64, 256),
  list_fail = T,
  dir = getwd(),
  attempt = 5,
  worker = 1
)
}
\arguments{
\item{time}{Date/time to download}

\item{type}{Type of lighting data to be loaded. Accepts "cc" and "cg"}

\item{range}{Radar range in km. Accepts 64 and 256.}

\item{list_fail}{List failed-to-download items}

\item{dir}{Directory of downloaded data.}

\item{attempt}{Attempts to be made per file to download}

\item{worker}{Numbers of sessions to be open}
}
\value{

}
\description{
HK weather - Downloading lighting & radar images from https://www.hko.gov.hk/en/index.html
}
\examples{
hk_load_ltng()
}
