% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk_load_sart_csv.R
\name{hk_load_sart_csv}
\alias{hk_load_sart_csv}
\title{HK weather - Downloading solar radiation information from https://data.gov.hk/en/help/api-spec}
\usage{
hk_load_sart_csv(
  time = weather2::tool_datetime(end = Sys.time(), by = "10 min", duration = "7 day"),
  lan = "en",
  list_fail = T,
  dir = getwd(),
  attempt = 5,
  worker = 1
)
}
\arguments{
\item{time}{Date/time to download}

\item{lan}{Language to download. Accepts "en", "tc", or "sc"}

\item{list_fail}{List failed-to-download items}

\item{dir}{Directory of downloaded data.}

\item{attempt}{Attempts to be made per file to download}

\item{worker}{Numbers of sessions to be open}
}
\value{

}
\description{
HK weather - Downloading solar radiation information from https://data.gov.hk/en/help/api-spec
}
\examples{
hk_load_sart_csv()
}
