% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk_load_satl.R
\name{hk_load_satl}
\alias{hk_load_satl}
\title{HK weather - Downloading satellite images from https://www.hko.gov.hk/en/index.html}
\usage{
hk_load_satl(
  time = weather2::tool_datetime(end = Sys.time(), by = "10 min", duration = "3 day"),
  magn = c(2, 4, 8),
  type = c("tc", "ir", "dc"),
  list_fail = T,
  dir = getwd(),
  attempt = 5,
  worker = 1
)
}
\arguments{
\item{time}{Date/time to download}

\item{magn}{Satellite image magnification. Accepts 2/ 4/ 8, or any combinations}

\item{type}{Type of satellite image. Accepts "tc" (True color), "ir" (Infrared), "dc" (Deep convection), or any combinations.}

\item{list_fail}{List failed-to-download items}

\item{dir}{Directory of downloaded data.}

\item{attempt}{Attempts to be made per file to download.}

\item{worker}{Numbers of sessions to be open}
}
\value{

}
\description{
HK weather - Downloading satellite images from https://www.hko.gov.hk/en/index.html
}
\examples{
hk_load_satl()
}
