% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk_load_tide_predict.R
\name{hk_load_tide_predict}
\alias{hk_load_tide_predict}
\title{HK weather - Download predicted tidal height data from https://www.hko.gov.hk/en/tide/ttext.htm}
\usage{
hk_load_tide_predict(station, year = 2015L:2025L, mode = c("hr", "hl"))
}
\arguments{
\item{station}{Station to download. Check the \code{hk_dict_tide} for more information}

\item{year}{Year to download}

\item{mode}{}
}
\description{
HK weather - Download predicted tidal height data from https://www.hko.gov.hk/en/tide/ttext.htm
}
\examples{
hk_load_tide_predict("CLK", c(2016L, 2021L))
}
