% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk_load_tide_predict.R
\name{hk_load_tide_predict_hl2015}
\alias{hk_load_tide_predict_hl2015}
\title{Download high/ low predicted tidal height data, using the year 2015 method}
\usage{
hk_load_tide_predict_hl2015(station, year)
}
\arguments{
\item{year}{}
}
\description{
Download high/ low predicted tidal height data, using the year 2015 method
}
\keyword{internal}
