% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk_load_tide_predict.R
\name{hk_load_tide_predict_hl2020}
\alias{hk_load_tide_predict_hl2020}
\title{Download high/ low predicted tidal height data, using the year 2020 method}
\usage{
hk_load_tide_predict_hl2020(station, year)
}
\arguments{
\item{station}{Station to download. Check the \code{hk_dict_tide} for more information}

\item{year}{Year to download}
}
\description{
Download high/ low predicted tidal height data, using the year 2020 method
}
\examples{
N/A
}
\keyword{internal}
