% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk_load_tide_predict.R
\name{hk_load_tide_predict_hr2015}
\alias{hk_load_tide_predict_hr2015}
\title{Download hourly predicted tidal height data, using the year 2015 method}
\usage{
hk_load_tide_predict_hr2015(station, year)
}
\arguments{
\item{station}{Station to download. Check the \code{hk_dict_tide} for more information}

\item{year}{Year to download}
}
\description{
Download hourly predicted tidal height data, using the year 2015 method
}
\examples{
N/A
}
\keyword{internal}
