% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk_load_tide_predict.R
\name{hk_load_tide_predict_hr2020}
\alias{hk_load_tide_predict_hr2020}
\title{Download hourly predicted tidal height data, using the year 2020 method}
\usage{
hk_load_tide_predict_hr2020(station, year)
}
\arguments{
\item{year}{}
}
\description{
Download hourly predicted tidal height data, using the year 2020 method
}
\keyword{internal}
