% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk_load_wcht.R
\name{hk_load_wcht}
\alias{hk_load_wcht}
\title{HK weather - Downloading weather chart from https://www.hko.gov.hk/en/index.html}
\usage{
hk_load_wcht(
  time = weather2::tool_datetime(end = Sys.time(), by = "1 hour", duration = "7 day"),
  list_fail = T,
  dir = getwd(),
  attempt = 5,
  worker = 1,
  threshold = 0.6
)
}
\arguments{
\item{time}{Date/time to download}

\item{list_fail}{List failed-to-download items}

\item{dir}{Directory of downloaded data.}

\item{attempt}{Attempts to be made per file to download}

\item{worker}{Numbers of sessions to be open}

\item{threshold}{Threshold of file size, when compared to the largest item to be installed in the list}
}
\value{

}
\description{
HK weather - Downloading weather chart from https://www.hko.gov.hk/en/index.html
}
\examples{
hk_load_wcht()
}
