% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk_load_wind.R
\name{hk_load_wind}
\alias{hk_load_wind}
\title{HK weather - Downloading wind distribution images from https://www.hko.gov.hk/en/index.html}
\usage{
hk_load_wind(
  lan = "en",
  type = c("wind", "gust"),
  list_fail = T,
  dir = getwd(),
  attempt = 5,
  worker = 1
)
}
\arguments{
\item{lan}{Language. Accepts "en", "tc", "sc".}

\item{type}{Type of satellite image. Accepts "tc" (True color), "ir" (Infrared), "dc" (Deep convection), or any combinations.}

\item{list_fail}{List failed-to-download items}

\item{dir}{Directory of downloaded data.}

\item{attempt}{Attempts to be made per file to download.}

\item{worker}{Numbers of sessions to be open}
}
\value{

}
\description{
HK weather - Downloading wind distribution images from https://www.hko.gov.hk/en/index.html
}
\examples{
hk_load_wind()
}
