% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk_plot_tide.R
\name{hk_plot_tide}
\alias{hk_plot_tide}
\title{Plots HK predicted tide, based on weeks}
\usage{
hk_plot_tide(station, year, weeks, range = NULL, draw_rect = T)
}
\arguments{
\item{station}{Station to download. Check the \code{hk_dict_tide} for more information}

\item{year}{Year to download, in \code{integer}}

\item{weeks}{Week to display, in \code{integer}}

\item{range}{Range of the appropriate tidal level. Default (\code{NULL}) assumes that any tidal level will be appropriate.}

\item{draw_rect}{Draw a shading rectangle between 18:00 to 06:00 next day to represent night time? Default as \code{TRUE}}
}
\description{
A common function used in SKLMP
}
\examples{
hk_plot_tide("CLK", 2020L, weeks = 10L:20L, range = c(0, 1), draw_rect = T)
}
