% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk_read_mslp.R
\name{hk_read_mslp}
\alias{hk_read_mslp}
\title{HK weather - Read the mslp files}
\usage{
hk_read_mslp(file_list)
}
\arguments{
\item{file_list}{a list of csv file directories}
}
\value{

}
\description{
HK weather - Read the mslp files
}
\examples{
hk_read_mslp(c("C:/Users/carol/Desktop/1.csv", "C:/Users/carol/Desktop/2.csv", "C:/Users/carol/Desktop/3.csv"))
}
