% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk_read_rain.R
\name{hk_read_rain}
\alias{hk_read_rain}
\title{HK weather - Read the RAIN files}
\usage{
hk_read_rain(file_list)
}
\arguments{
\item{file_list}{a list of csv file directories}
}
\value{

}
\description{
HK weather - Read the RAIN files
}
\examples{
hk_read_rain(c("C:/Users/carol/Desktop/1.csv", "C:/Users/carol/Desktop/2.csv", "C:/Users/carol/Desktop/3.csv"))
}
