% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk_read_rhum.R
\name{hk_read_rhum}
\alias{hk_read_rhum}
\title{HK weather - Read the GTMP files}
\usage{
hk_read_rhum(file_list)
}
\arguments{
\item{file_list}{a list of csv file directories}
}
\value{

}
\description{
HK weather - Read the GTMP files
}
\examples{
hk_read_rhum(c("C:/Users/carol/Desktop/1.csv", "C:/Users/carol/Desktop/2.csv", "C:/Users/carol/Desktop/3.csv"))
}
