% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk_read_sart.R
\name{hk_read_sart}
\alias{hk_read_sart}
\title{HK weather - Read the SART files}
\usage{
hk_read_sart(file_list)
}
\arguments{
\item{file_list}{a list of csv file directories}
}
\value{

}
\description{
HK weather - Read the SART files
}
\examples{
hk_read_sart(c("C:/Users/carol/Desktop/1.csv", "C:/Users/carol/Desktop/2.csv", "C:/Users/carol/Desktop/3.csv"))
}
