% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk_read_temp.R
\name{hk_read_temp}
\alias{hk_read_temp}
\title{HK weather - Read the TEMP files}
\usage{
hk_read_temp(file_list)
}
\arguments{
\item{file_list}{a list of csv file directories}
}
\value{

}
\description{
HK weather - Read the TEMP files
}
\examples{
hk_read_temp(c("C:/Users/carol/Desktop/1.csv", "C:/Users/carol/Desktop/2.csv", "C:/Users/carol/Desktop/3.csv"))
}
