% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk_read_tide.R
\name{hk_read_tide}
\alias{hk_read_tide}
\title{HK weather - Read the tide files}
\usage{
hk_read_tide(file_list)
}
\arguments{
\item{file_list}{a list of csv file directories}
}
\value{

}
\description{
HK weather - Read the tide files
}
\examples{
hk_read_tide(c("C:/Users/carol/Desktop/1.csv", "C:/Users/carol/Desktop/2.csv", "C:/Users/carol/Desktop/3.csv"))
}
