% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_ckf_HKLoad.R
\name{sys_ckf_HKLoad}
\alias{sys_ckf_HKLoad}
\title{Check if items within the HKLoad or MOLoad are valid}
\usage{
sys_ckf_HKLoad(time, list_fail, attempt, worker)
}
\arguments{
\item{time}{time within the function}

\item{list_fail}{list_fail within the function}

\item{attempt}{attempt within the function}

\item{worker}{worker within the function}
}
\value{

}
\description{
Check if items within the HKLoad or MOLoad are valid
}
\examples{
sys_ckf_HKLoad(time, list_fail, attempt, worker)
}
