% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_ckf_HKLoad.R
\name{sys_ckf_HKLoadLan}
\alias{sys_ckf_HKLoadLan}
\title{Check if language is supported}
\usage{
sys_ckf_HKLoadLan(lan)
}
\arguments{
\item{lan}{lan within the function}
}
\value{

}
\description{
Check if language is supported
}
